change tests to json; sort output
