+++ kani-compiler/src/kani_middle/codegen_units.rs
@@ -9,8 +9,7 @@
 
 use crate::args::ReachabilityType;
 use crate::kani_middle::attributes::is_proof_harness;
-use crate::kani_middle::list::collect_contracted_fns;
-use crate::kani_middle::metadata::gen_proof_metadata;
+use crate::kani_middle::metadata::{gen_proof_metadata, gen_contracts_metadata};
 use crate::kani_middle::reachability::filter_crate_items;
 use crate::kani_middle::resolve::expect_resolve_fn;
 use crate::kani_middle::stubbing::{check_compatibility, harness_stub_map};
@@ -115,7 +114,7 @@ impl CodegenUnits {
             proof_harnesses,
             unsupported_features: vec![],
             test_harnesses,
-            contracted_functions: collect_contracted_fns(tcx),
+            contracted_functions: gen_contracts_metadata(tcx),
         }
     }
 }
