+++ tests/script-based-pre/cargo_list/Cargo.toml
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0 OR MIT
 
 [package]
-name = "sample_crate"
+name = "cargo_list"
 version = "0.1.0"
 edition = "2021"
