+++ tests/script-based-pre/cargo_list/config.yml
@@ -1,4 +1,4 @@
 # Copyright Kani Contributors
 # SPDX-License-Identifier: Apache-2.0 OR MIT
 script: list.sh
-expected: list-pretty.expected
+expected: list.expected
