--- tests/script-based-pre/cargo_list/list-pretty.expected
@@ -1,36 +1,0 @@
-Contracts:
-Each function in the table below either has contracts or is the target of a contract harness (#[kani::proof_for_contract]).
-
-Function
-# of Contracts
-Contract Harnesses     
-
-example::implementation::bar
-4
-example::verify::check_bar     
-
-example::implementation::baz
-0
-example::verify::check_baz   
-
-example::implementation::foo
-2
-example::verify::check_foo_u32
-
-example::verify::check_foo_u64 
-example::implementation::func
-1
-example::verify::check_func    
-
-example::prep::parse
-1
-NONE   
-
-Total
-5
-8
-5
-
-Standard Harnesses (#[kani::proof]):
-1. standard_harnesses::example::verify::check_modify
-2. standard_harnesses::example::verify::check_new
