+++ tests/script-based-pre/cargo_list/list.sh
@@ -2,7 +2,9 @@
 # Copyright Kani Contributors
 # SPDX-License-Identifier: Apache-2.0 OR MIT
 
-# The table in the expected omits the table borders because the runtest script
-# does not evaluate the table borders in the captured output as equal to the table borders in the expected file.
+# Check that the JSON file produced by `cargo kani list` is correct.
+# Note that the list.expected file omits the value for "kani-version"
+# to avoid having to update the test every time we bump versions.
 
-cargo kani list -Z list
+cargo kani list -Z list -Z function-contracts --format json
+cat "kani-list.json"
